Resolve: Update nitrite to 4.2.1.

Closes #84 and #87

See merge request bright-giant/sirius/sirius-libs!53

+++ chemical_db_oss/chemdb_sql_jdbi_json_oss/src/main/java/de/unijena/bioinf/chemdb/SpectralSqlJdbiJsonDatabase.java
@@ -56,7 +56,7 @@ public class SpectralSqlJdbiJsonDatabase implements SpectralLibrary {
     }
 
     @Override
-    public int countAllSpectra() {
+    public long countAllSpectra() {
         return jdbi.withExtension(Dao.class, Dao::countAll);
     }
